Compile after Math3d was changed to use float instead of qreal.
Introduced by qtbase:
51d40d7e9bdfc63c5109aef5b732aa2ba10f985a .
Fix warnings about truncation from double to float.

Introduce qFuzzyCompare() and mark some tests as
insignificant.

Task-number: QTBUG-27189

+++ src/threed/geometry/qglsection.cpp
@@ -140,7 +140,7 @@ static inline bool qSameDirection(const QVector3D &a , const QVector3D &b)
     if (!a.isNull() && !b.isNull())
     {
         float dot = QVector3D::dotProduct(a, b);
-        res = qFskCompare((qreal)dot, a.length() * b.length());
+        res = qFskCompare(dot, a.length() * b.length());
     }
     return res;
 }
